libcrux_hacl_rs::bignum::bignum256

Function mod_inv_prime_vartime

pub fn mod_inv_prime_vartime(n: &[u64], a: &[u64], res: &mut [u64]) -> bool

Write a ^ (-1) mod n in res.

The arguments a, n and the outparam res are meant to be 256-bit bignums, i.e. uint64_t\[4\].

Before calling this function, the caller will need to ensure that the following preconditions are observed. • n is a prime

The function returns false if any of the following preconditions are violated, true otherwise. • n % 2 = 1 • 1 < n • 0 < a • a < n