libcrux_hacl_rs::bignum::bignum256

Function new_bn_from_bytes_be

pub fn new_bn_from_bytes_be(len: u32, b: &[u8]) -> Box<[u64]>

Load a bid-endian bignum from memory.

The argument b points to len bytes of valid memory. The function returns a heap-allocated bignum of size sufficient to hold the result of loading b, or NULL if either the allocation failed, or the amount of required memory would exceed 4GB.

If the return value is non-null, clients must eventually call free(3) on it to avoid memory leaks.