libcrux_hacl_rs::bignum::bignum32

Function mod_inv_prime_vartime

pub fn mod_inv_prime_vartime(
    len: u32,
    n: &[u32],
    a: &[u32],
    res: &mut [u32],
) -> bool

Write a ^ (-1) mod n in res.

input a: Points to len number of limbs, i.e. uint32_t\[len\]. Must be disjoint from the memory locations of n and res.

input n: Points to len number of limbs, i.e. uint32_t\[len\]. Must be disjoint from the memory locations of a and res.

output res: Points to len number of limbs, i.e. uint32_t\[len\]. Must be disjoint from the memory location of a and n.

@return false if any preconditions (except the precondition: n is a prime) are violated, true otherwise.

@pre Before calling this function, the caller will need to ensure that the following preconditions are observed: - n is a prime - n % 2 = 1 - 1 < n - 0 < a - a < n